+++ packages/types/src/types/cloudinary-video-player.ts
@@ -302,6 +302,13 @@ export interface CloudinaryVideoPlayer {
    * @see https://cloudinary.com/documentation/video_player_api_reference#fluid
    */
   fluid: (fluid?: boolean) => boolean;
+
+  videojs: {
+    cloudinary: {
+      dispose: () => void;
+    }
+  }
+
 }
 
 export interface CloudinaryVideoPlayerOptionsLogo {
@@ -435,6 +442,7 @@ export interface CloudinaryVideoPlayerOptions
   };
   analytics?: boolean;
   allowUsageReport?: boolean;
+  cloudinaryAnalytics?: boolean | CloudinaryVideoPlayerOptionsCloudinaryAnalytics;
 
   // ------------ Delivery ------------
   cloud_name?: string;
@@ -475,4 +483,16 @@ export interface CloudinaryVideoPlayerOptionPosterOptions {
    * RGBA hex, or a named color.
    */
   posterColor?: string;
+}
+
+export interface CloudinaryVideoPlayerOptionsCloudinaryAnalytics {
+  customData: CloudinaryVideoPlayerOptionsCloudinaryAnalyticsCustomData;
+}
+
+export interface CloudinaryVideoPlayerOptionsCloudinaryAnalyticsCustomData {
+  customData1?: string;
+  customData2?: string;
+  customData3?: string;
+  customData4?: string;
+  customData5?: string;
 }